+++ src/git.rs
@@ -1,19 +1,12 @@
 use std::process::Command;
 
 pub fn branch() -> Option<String> {
-
     let output = Command::new("git")
-        .args([
-            "branch",
-            "--show-current"
-        ])
+        .args(["branch", "--show-current"])
         .output()
         .ok()?;
 
-    let branch =
-        String::from_utf8_lossy(&output.stdout)
-        .trim()
-        .to_string();
+    let branch = String::from_utf8_lossy(&output.stdout).trim().to_string();
 
     if branch.is_empty() {
         None